refactor: rearrange car max name feature to Car
